\name{calcDiff}
\alias{calcDiff}
\title{Calculate difference between two numeric vectors that
gives exactly zero for very small relative differences.}
\usage{
calcDiff(X, Y, EPS = 1e-12)
}
\arguments{
  \item{X}{first number}

  \item{Y}{second number}

  \item{EPS}{relative difference equivalent to zero}
}
\value{
difference
}
\description{
Calculate difference between two numeric vectors that gives
exactly zero for very small relative differences.
}

