 \name{creatinine}
     \docType{data}
     \alias{creatinine}
     \title{Comparison of blood and serum creatinine measurement}
     \description{
       This data set gives the blood and serum preoperative creatinine
       measurements in 110 heart surgery patients.
     }
     \usage{data(creatinine)}
     \format{A data frame containing 110 observations with serum and plasma creatinin measurements in mg/dL for each sample.}
     \keyword{datasets}
