% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_mcr.r
\name{get_e0_lm}
\alias{get_e0_lm}
\title{Get standard ("original") mean-squared-error for linear or kernel models.}
\usage{
get_e0_lm(model, suff_stats)
}
\arguments{
\item{model}{a coefficient vector (see \code{\link{get_lm_regularized}})}

\item{suff_stats}{output from \code{\link{get_suff_stats_kernel}}, or from the internal function \code{get_suff_stats_lm}. See also \code{\link{precompute_mcr_objects_and_functions}}}
}
\value{
Mean-squared-error in the sample.
}
\description{
Can be used to calculate loss of a reference model, to determine epsilon.
}
