% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge_mcr.r
\name{hinge_set}
\alias{hinge_set}
\alias{hinge_pred}
\alias{hinge_set}
\alias{hinge_w}
\title{Calculate the hinge loss}
\usage{
hinge_pred(pred, y, m = 1)

hinge_w(w, X, ...)
}
\arguments{
\item{pred}{y predictions from the model w}

\item{y}{outcome vector, with elements -1 or 1}

\item{m}{margin parameter}

\item{w}{weight vector (linear classification model). First element should be an interecept term. For example, the output of \code{\link{optimize_hinge_general}}.}

\item{X}{covariate matrix, should not include a constant (intercept) column.}

\item{...}{passed to hinge_pred}
}
\description{
\code{hinge_w} computes predictions, and passes these to \code{hinge_pred} to compute the in-sample hinge loss.
}
