\name{MCResult.getCoefficients}
\alias{getCoefficients}
\alias{MCResult.getCoefficients}
\title{Get Regression Coefficients}
\usage{
  MCResult.getCoefficients(.Object)
}
\arguments{
  \item{.Object}{object of class "MCResult".}
}
\value{
  Regression parameters in matrix form. Rows: Intercept,
  Slope. Cols: EST, SE, LCI, UCI.
}
\description{
  Get Regression Coefficients
}

