\name{MCResult.getRegmethod}
\alias{MCResult.getRegmethod}
\alias{getRegmethod}
\title{Get Regression Method}
\usage{
MCResult.getRegmethod(.Object)
}
\arguments{
  \item{.Object}{object of class "MCResult".}
}
\value{
Name of the statistical method used for the regression
analysis.
}
\description{
Get Regression Method
}

