\name{MCResult.getWeights}
\alias{getWeights}
\alias{MCResult.getWeights}
\title{Get Weights of Data Points}
\usage{
  MCResult.getWeights(.Object)
}
\arguments{
  \item{.Object}{Object of class "MCResult"}
}
\value{
  Weights of data points.
}
\description{
  Get Weights of Data Points
}

