\name{MCResult.printSummary}
\alias{MCResult.printSummary}
\alias{printSummary}
\title{Print Summary of a Regression Analysis}
\usage{
MCResult.printSummary(.Object)
}
\arguments{
  \item{.Object}{object of type "MCResult".}
}
\description{
Print Summary of a Regression Analysis
}
\seealso{
\code{\link{getCoefficients}}, \code{\link{getRegmethod}}
}

