\name{MCResultBCa.bootstrapSummary}
\alias{MCResultBCa.bootstrapSummary}
\title{Compute Bootstrap-Summary for 'MCResultBCa' Objects.}
\usage{
MCResultBCa.bootstrapSummary(.Object)
}
\arguments{
  \item{.Object}{object of class 'MCResultBCa'}
}
\value{
matrix of bootstrap results
}
\description{
Function computes the bootstrap summary for objects of
class 'MCResultBCa'.
}

