% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultBCaMethods.r
\name{MCResultBCa.plotBootstrapCoefficients}
\alias{MCResultBCa.plotBootstrapCoefficients}
\title{Plot distriblution of bootstrap coefficients}
\usage{
MCResultBCa.plotBootstrapCoefficients(.Object, breaks = 20, ...)
}
\arguments{
\item{.Object}{Object of class "MCResultBCa"}

\item{breaks}{used in function 'hist' (see ?hist)}

\item{...}{further graphical parameters}
}
\value{
No return value
}
\description{
Plot distriblution of bootstrap coefficients (slope and intercept).
}
