% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultBCaMethods.r
\name{MCResultBCa.plotBootstrapT}
\alias{MCResultBCa.plotBootstrapT}
\title{Plot distriblution of bootstrap pivot T}
\usage{
MCResultBCa.plotBootstrapT(.Object, breaks = 20, ...)
}
\arguments{
\item{.Object}{Object of class "MCResultBCa".}

\item{breaks}{Number of breaks in histogram.}

\item{...}{further graphical parameters}
}
\value{
No return value
}
\description{
Plot distriblution of bootstrap pivot T for slope and intercept and compare
them with t(n-2) distribution.
}
