\name{MCResultBCa.printSummary}
\alias{MCResultBCa.printSummary}
\title{Print Regression-Analysis Summary for Objects of class 'MCResultBCa'.}
\usage{
MCResultBCa.printSummary(.Object)
}
\arguments{
  \item{.Object}{object of class 'MCResultBCa'}
}
\description{
Functions prints a summary of the regression-analysis for
objects of class 'MCResultBCa'.
}

