% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultJackknifeMethods.r
\name{MCResultJackknife.calcResponse}
\alias{MCResultJackknife.calcResponse}
\title{Caluculate Response}
\usage{
MCResultJackknife.calcResponse(.Object, x.levels, alpha = 0.05)
}
\arguments{
\item{.Object}{object of class 'MCResultJackknife'}

\item{x.levels}{numeric vector specifying values of the reference method for which prediction should be made}

\item{alpha}{significance level for confidence intervals}
}
\value{
matrix with predicted values with confidence intervals for given values of the reference-method.
}
\description{
Calculate predicted values for given values of the reference-method.
}
