\name{MCResultJackknife.getJackknifeIntercept}
\alias{MCResultJackknife.getJackknifeIntercept}
\title{Get-Method for Jackknife-Intercept Value.}
\usage{
MCResultJackknife.getJackknifeIntercept(.Object)
}
\arguments{
  \item{.Object}{object of class 'MCResultJackknife'}
}
\value{
(numeric) jackknife-intercept
}
\description{
Extracts the intercept value from objects of class
'MCResultJackknife'.
}

