\name{MCResultJackknife.getJackknifeSlope}
\alias{MCResultJackknife.getJackknifeSlope}
\title{Get-Method for Jackknife-Slope Value.}
\usage{
MCResultJackknife.getJackknifeSlope(.Object)
}
\arguments{
  \item{.Object}{object of class 'MCResultJackknife'}
}
\value{
(numeric) jackknife-slope
}
\description{
Extracts the slope value from objects of class
'MCResultJackknife'.
}

