\name{MCResultJackknife.printSummary}
\alias{MCResultJackknife.printSummary}
\title{Print Regression-Analysis Summary for Objects of class 'MCResultJackknife'.}
\usage{
MCResultJackknife.printSummary(.Object)
}
\arguments{
  \item{.Object}{object of class 'MCResultJackknife'}
}
\description{
Functions prints a summary of the regression-analysis for
objects of class 'MCResultJackknife'.
}

