\name{MCResultResampling.bootstrapSummary}
\alias{MCResultResampling.bootstrapSummary}
\title{Compute Bootstrap-Summary for 'MCResultResampling' Objects.}
\usage{
MCResultResampling.bootstrapSummary(.Object)
}
\arguments{
  \item{.Object}{object of class 'MCResultResampling'}
}
\value{
matrix of bootstrap results
}
\description{
Function computes the bootstrap summary for objects of
class 'MCResultResampling'.
}

