% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultResamplingMethods.r
\name{MCResultResampling.plotBootstrapCoefficients}
\alias{MCResultResampling.plotBootstrapCoefficients}
\title{Plot distriblution of bootstrap coefficients}
\usage{
MCResultResampling.plotBootstrapCoefficients(.Object, breaks = 20, ...)
}
\arguments{
\item{.Object}{Object of class "MCResultResampling"}

\item{breaks}{see function 'hist' (?hist) for details}

\item{...}{further graphical parameters}
}
\value{
No return value
}
\description{
Plot distriblution of bootstrap coefficients (slope and intercept).
}
