% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_mcr.r
\name{norm_RKHS}
\alias{norm_RKHS}
\title{Get function norm for regression models in a RKHS}
\usage{
norm_RKHS(K_D = NA, dat_ref = NA, model, kern_fun)
}
\arguments{
\item{K_D}{kernel function applied to the reference data, in order to obtain a regularization matrix.}

\item{dat_ref}{(optional) used to calculate \code{K_D}, if not specified.}

\item{model}{model vector for which to calculate the norm.}

\item{kern_fun}{(optional) used to calculate \code{K_D}, if not specified.}
}
\value{
norm of \code{model} with respect to the regularization matrix \code{K_D}. This can be used in defining a model class, i.e., as the set of all models will norm less than that of a reference model.
}
\description{
Get function norm for regression models in a RKHS
}
