% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_search_mcr.r
\name{plot_empirical_MCR_searchpath}
\alias{plot_empirical_MCR_searchpath}
\title{Plot all models found during MCR binary search, and the associated MR boundaries for any model in the class.}
\usage{
plot_empirical_MCR_searchpath(emp_mcr, eps = NA, ylab = NA,
  xlab = NA, show_all_lines = FALSE, show_mcr_range = FALSE,
  add_to_plot = FALSE, ylim_include = NULL, xlim_include = NULL,
  xlim_1 = TRUE, fill_col = NA, ...)
}
\arguments{
\item{emp_mcr}{output from `get_empirical_MCR`}

\item{eps}{absolute error value, around which empirical MCR is defined}

\item{ylab, xlab}{axis labels, as in \code{\link{plot}}}

\item{show_all_lines}{show redundant MR boundaries found in MCR search}

\item{show_mcr_range}{additionally plots brackets corresponding to empirical MCR}

\item{add_to_plot}{whether to add lines to an existing plot, or to create a new plot}

\item{ylim_include}{vector of numeric values that should be contained in ylim}

\item{xlim_include}{vector of numeric values that should be contained in xlim}

\item{xlim_1}{whether xlim should truncate at value 1 (null reliance)}

\item{fill_col}{color to shade in region below permorfance threshold. If NA, no shading is applied.}

\item{...}{passed to \code{\link{points}} & \code{\link{matlines}}.}
}
\description{
Plot all models found during MCR binary search, and the associated MR boundaries for any model in the class.
}
