% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultMethods.r
\name{MCResult.getCoefficients}
\alias{MCResult.getCoefficients}
\alias{getCoefficients}
\alias{coef}
\title{Get Regression Coefficients}
\usage{
MCResult.getCoefficients(.Object)
}
\arguments{
\item{.Object}{object of class "MCResult".}
}
\value{
Regression parameters in matrix form. Rows: Intercept, Slope. Cols: EST, SE, LCI, UCI.
}
\description{
Get Regression Coefficients
}
