% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultMethods.r
\name{MCResult.getData}
\alias{MCResult.getData}
\alias{getData}
\title{Get Data}
\usage{
MCResult.getData(.Object)
}
\arguments{
\item{.Object}{object of class "MCResult".}
}
\value{
Measurement data in matrix format. First column contains reference method (X), second column contains test method (Y).
}
\description{
Get Data
}
