\name{MCResult.getErrorRatio}
\alias{MCResult.getErrorRatio}
\alias{getErrorRatio}
\title{Get Error Ratio}
\usage{
MCResult.getErrorRatio(.Object)
}
\arguments{
  \item{.Object}{Object of class "MCResult"}
}
\value{
Error ratio. Only relevant for Deming type regressions.
}
\description{
Get Error Ratio
}

