% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultMethods.r
\name{MCResult.getWeights}
\alias{MCResult.getWeights}
\alias{getWeights}
\title{Get Weights of Data Points}
\usage{
MCResult.getWeights(.Object)
}
\arguments{
\item{.Object}{Object of class "MCResult"}
}
\value{
Weights of data points.
}
\description{
Get Weights of Data Points
}
