% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultMethods.r
\name{MCResult.printSummary}
\alias{MCResult.printSummary}
\alias{printSummary}
\alias{summary}
\title{Print Summary of a Regression Analysis}
\usage{
MCResult.printSummary(.Object)
}
\arguments{
\item{.Object}{object of type "MCResult".}
}
\value{
No return value
}
\description{
Print Summary of a Regression Analysis
}
\seealso{
\code{\link{getCoefficients}}, \code{\link{getRegmethod}}
}
