% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultAnalyticalMethods.r
\name{MCResultAnalytical.printSummary}
\alias{MCResultAnalytical.printSummary}
\title{Print Regression-Analysis Summary for Objects of class 'MCResultAnalytical'.}
\usage{
MCResultAnalytical.printSummary(.Object)
}
\arguments{
\item{.Object}{object of class 'MCResultAnalytical'}
}
\value{
No return value
}
\description{
Function prints a summary of the regression-analysis for objects of class 'MCResultAnalytical'.
}
