% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultBCaMethods.r
\name{MCResultBCa.printSummary}
\alias{MCResultBCa.printSummary}
\title{Print Regression-Analysis Summary for Objects of class 'MCResultBCa'.}
\usage{
MCResultBCa.printSummary(.Object)
}
\arguments{
\item{.Object}{object of class 'MCResultBCa'}
}
\value{
No return value
}
\description{
Functions prints a summary of the regression-analysis for objects of class 'MCResultBCa'.
}
