\name{MCResultJackknife.calcResponse}
\alias{MCResultJackknife.calcResponse}
\title{Caluculate Response}
\usage{
MCResultJackknife.calcResponse(.Object, x.levels, alpha = 0.05)
}
\arguments{
  \item{.Object}{object of class 'MCResultJackknife'}

  \item{x.levels}{numeric vector specifying values of the
  reference method for which prediction should be made}

  \item{alpha}{significance level for confidence intervals}
}
\description{
Calculate predicted values for given values of the
reference-method.
}

