% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCResultJackknifeMethods.r
\name{MCResultJackknife.getJackknifeStatistics}
\alias{MCResultJackknife.getJackknifeStatistics}
\title{Jackknife Statistics}
\usage{
MCResultJackknife.getJackknifeStatistics(.Object)
}
\arguments{
\item{.Object}{object of class "MCResultJackknife" or "MCResultResampling"}
}
\value{
table with jackknife mean, bias and standard error for intercept and slope.
}
\description{
Calculate jackknife mean, bias and standard error.
}
