\name{MCResultResampling.plotBootstrapCoefficients}
\alias{MCResultResampling.plotBootstrapCoefficients}
\title{Plot distriblution of bootstrap coefficients}
\usage{
MCResultResampling.plotBootstrapCoefficients(.Object, breaks = 20, ...)
}
\arguments{
  \item{.Object}{Object of class "MCResultResampling"}

  \item{breaks}{see function 'hist' (?hist) for details}

  \item{...}{further graphical parameters}
}
\description{
Plot distriblution of bootstrap coefficients (slope and
intercept).
}

