\name{MCResultResampling.printSummary}
\alias{MCResultResampling.printSummary}
\title{Print Regression-Analysis Summary for Objects of class 'MCResultResampling'.}
\usage{
MCResultResampling.printSummary(.Object)
}
\arguments{
  \item{.Object}{object of class 'MCResultResampling'}
}
\description{
Functions prints a summary of the regression-analysis for
objects of class 'MCResultResampling'.
}

